% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{lsystem}
\alias{lsystem}
\title{L-system}
\usage{
lsystem(w, p, r, n)
}
\arguments{
\item{w}{character}

\item{p}{pattern}

\item{r}{replacement}

\item{n}{number of loop}
}
\description{
L-system
}
\examples{
### Model of the growth of algae
0:6 |> lapply(\(x) lsystem("A", c("A", "B"), c("AB", "A"), x)) |> unlist()

}
