% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{lsystem2}
\alias{lsystem2}
\title{L-system version 2}
\usage{
lsystem2(w, pr, n)
}
\arguments{
\item{w}{character}

\item{pr}{replacement rule}

\item{n}{number of loop}
}
\description{
L-system version 2
}
\examples{
### Model of the growth of algae
0:6 |> lapply(\(x) lsystem2("A", c("A" = "AB", "B" = "A"), x)) |> unlist()

### Fibonacci numbers
0:6 |> lapply(\(x) lsystem2("A", c("A" = "B", "B" = "AB"), x)) |> unlist()

### Cantor set
0:4 |> lapply(\(x) lsystem2("A", c("A" = "ABA", "B" = "BBB"), x)) |> unlist()

### Koch curve
0:4 |> lapply(\(x) lsystem2("F", c("F" = "F+F-F-F+F"), x)) |> unlist()

}
