% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot tibble curve data}
\usage{
plot_curve(tbl_curve)
}
\arguments{
\item{tbl_curve}{tibble output from w2curve()}
}
\description{
Plot tibble curve data
}
\examples{
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |>
  w2curve2() |>
  plot_curve()

}
