% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{w2curve}
\alias{w2curve}
\title{Convert L-system symbol to data which able to plot by ggplot2}
\usage{
w2curve(w)
}
\arguments{
\item{w}{L-system symbol}
}
\description{
Convert L-system symbol to data which able to plot by ggplot2
}
\examples{
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve()

}
