% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{w2curve2}
\alias{w2curve2}
\title{w2curve() version 2}
\usage{
w2curve2(w)
}
\arguments{
\item{w}{L-system symbol}
}
\description{
w2curve() version 2
}
\examples{
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve2()

}
