% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{w2curve3}
\alias{w2curve3}
\title{w2curve() version 3 (with simple angle rule)}
\usage{
w2curve3(w, theta = pi/2)
}
\arguments{
\item{w}{L-system symbol}

\item{theta}{angle theta}
}
\description{
w2curve() version 3 (with simple angle rule)
}
\examples{
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve3()
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve3(theta = pi/3)

}
