% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsystem.R
\name{w2curve4}
\alias{w2curve4}
\title{w2curve() version 4 (with angle rule)}
\usage{
w2curve4(w, char_segment = "F", angle_rule = c(`+` = pi/2, `-` = -pi/2))
}
\arguments{
\item{w}{L-system symbol}

\item{char_segment}{specify segment character}

\item{angle_rule}{angle rule}
}
\description{
w2curve() version 4 (with angle rule)
}
\examples{
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve4()
lsystem2("F", c("F" = "F+F-F-F+F"), 4) |> w2curve4(angle_rule = c("+" = pi/3, "-" = pi/2))

}
